// es-ts/ts-nest-demo/src/sdks/general-test/dto/create.general.test.response.dto.ts
export interface StatefulTestResponse {
  id: string;
  studentKey: string;
  status: number;
  statusDetail: unknown;
  metaData: unknown;
  groupKey: string;
  product: number;
  productModule: number;
  businessKey: string;
  resultKey: string;
  result: unknown;
  createBy: string;
  updateBy: string;
  succeeded: boolean;
}